Extract reflection helpers into separate file and remove a number of warning suppressions.
